feat!: new toml format for buddies file (#17)
The default buddies file format is now `toml` the old `yaml` format is
deprecated and shouldn't be used anymore

BREAKING-CHANGE: `yaml` buddies configs have been deprecated in favor of
`toml`. For users using the default config locations the old config will
be automatically migrated to the new format. For users using custom
`--buddies-file` locations there is a new command `git-squad
migrate-buddies` that can be used to perform the migration